refactor(UI): Centralize config panel logic to fix button bug
Refactored the config category panel (`configCategoryPanel`) to use a single source of truth for generating its list of system buttons.

- Created a new `getVisibleConfigSystems` function in `core/ui/uiUtils.js`.
- Both `panelBuilder.js` and `panelHandlers.js` now call this function to get the list of systems.
- This permanently resolves the recurring "shifted buttons" bug where the UI and the handler would get out of sync.

feat(xray): Implement configurable X-ray ore detection system

Added a new feature to detect and notify staff about potential X-ray usage.

- Created `xrayConfig.js` to define monitored ores, dimensions, and Y-levels.
- Added `xrayDetection.js` module to monitor `playerBreakBlock` events.
- Implemented a new `/xraynotify` (alias `/xray`) command for staff to toggle notifications.
- Added a full set of UI panels under the "Detections" category for managing the X-ray system settings and monitored ores.

fix(UI): Address multiple UI bugs and improvements

- Added the missing "General Economy Settings" UI panel.
- Removed the misplaced "Bounties" button from the main economy panel.
- Set a default value for the rank nametag prefix to prevent errors with new ranks.
- Changed the color of the "Reset Settings" button to dark red (`§4`) for better visibility.
- Ensured all colored UI text is properly terminated with a reset code (`§r`).
