+++ AddonExeBP/scripts/core/ui/panelBuilder.js
@@ -14,6 +14,7 @@ import { items as allItems } from '../itemsConfig.js';
 import { getAllKits } from '../kitAdminManager.js';
 import { getValueFromPath } from '../objectUtils.js';
 import { formatCurrency } from '../utils.js';
+import { getVisibleConfigSystems } from './uiUtils.js';
 
 const itemsPerPage = 8;
 
@@ -1112,38 +1113,14 @@ export async function buildPanelForm(player, panelId, context) {
             const form = new ActionFormData().title(`${title} (Page ${page})`);
             form.button('§l§8< Back', 'textures/gui/controls/left.png');
 
-            let allSystems = [
-                ...configPanelSchema.filter(c => c.id !== 'economyGeneralSettings').map(c => ({ id: `config_${c.id}`, title: c.title, icon: c.icon }))
-            ];
-
-            if (pData.permissionLevel <= 1) {
-                allSystems.push({ id: 'kitManagementPanel', title: '§l§dKit System§r', icon: 'textures/ui/inventory_icon' });
-                allSystems.push({ id: 'shopManagementPanel', title: '§l§2Shop System§r', icon: 'textures/items/emerald' });
-                allSystems.push({ id: 'rankManagementPanel', title: '§l§4Rank System§r', icon: 'textures/ui/permissions_member_star.png' });
-                allSystems.push({ id: 'economyPanel', title: '§l§6Economy System§r', icon: 'textures/items/emerald' });
-            }
-            if (pData.permissionLevel === 0) {
-                allSystems.push({ id: 'configResetPanel', title: '§l§cReset Settings§r', icon: 'textures/ui/wysiwyg_reset' });
-            }
-
-            // Custom sorting: General first, Reset last, rest alphabetical
-            const generalSystem = allSystems.find(s => s.id === 'config_general');
-            const resetSystem = allSystems.find(s => s.id === 'configResetPanel');
-            let otherSystems = allSystems.filter(s => s.id !== 'config_general' && s.id !== 'configResetPanel');
-            otherSystems.sort((a, b) => a.title.replace(/§./g, '').localeCompare(b.title.replace(/§./g, '')));
-
-            const sortedSystems = [];
-            if (generalSystem) {sortedSystems.push(generalSystem);}
-            sortedSystems.push(...otherSystems);
-            if (resetSystem) {sortedSystems.push(resetSystem);}
-
+            const sortedSystems = getVisibleConfigSystems(pData);
             const paginatedSystems = getPaginatedItems(sortedSystems, page);
 
             for (const system of paginatedSystems) {
                 form.button(system.title, system.icon);
             }
 
-            addPaginationButtons(form, page, allSystems.length);
+            addPaginationButtons(form, page, sortedSystems.length);
             return form;
         }
